requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
jinja2==3.0.1
websockets==10.0
pydantic==1.8.2
redis

